docs/examples/cdk/deploy_to_ec2/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
strands-agents
strands-agents-tools
requests
requests-oauthlib
boto3
python-dotenv